Changed i32 to u32

+++ src/lib.rs
@@ -184,7 +184,7 @@ pub fn debug(msg: &str) {
   let mut mem_stack: SinglePageStack;
   unsafe { mem_stack = g_mem_stack.unwrap(); }
   let allocation_of_input =  serialize(&mut mem_stack, msg);
-  unsafe { hc_debug(allocation_of_input.encode() as i32); }
+  unsafe { hc_debug(allocation_of_input.encode()); }
   mem_stack.deallocate(allocation_of_input)
     .expect("should be able to deallocate input that has been allocated on memory stack");
 }
